backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
Django==3.2.20
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.100.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==0.21.1
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1